Create README.md
I'm a man of unlimited capabilities, (many yet to be explored) in a time where constant change requires quick whit,inventory, and maluabillity that is self taught, independent, in demand desired and most of all essential to manifest the conversion of the earth from it's present state so that all life forms can exist and transcend to a consciousness capable of intergalactic involvement, gratitude,and service to all that will ever be. Addicted to building greatness!